+++ src/Models/subscription/SubscriptionDiscount.ts
@@ -1,0 +1,7 @@
+import SubscriptionDiscountInterface from '../../Interfaces/subscription/SubscriptionDiscount';
+
+export default class SubscriptionDiscount {
+    constructor(parameters: SubscriptionDiscountInterface) {
+        Object.assign(this, parameters);
+    }
+}